[core] Renamed buffer.cpp to buffer_snd.cpp.
