+++ srtcore/buffer_snd.h
